Reorganise filenames
